requirements.txt
requests
boto3
textract-trp